+++ lua/lazy/docs.lua
@@ -131,7 +131,8 @@ function M.colors(opts)
     { "---", "---", "---" },
   }
   Util.foreach(require(opts.modname).colors, function(group, link)
-    lines[#lines + 1] = { "**" .. opts.name .. group .. "**", "***" .. link .. "***", comments[group] or "" }
+    link = type(link) == "table" and "`" .. vim.inspect(link):gsub("%s+", " ") .. "`" or "***" .. link .. "***"
+    lines[#lines + 1] = { "**" .. opts.name .. group .. "**", link, comments[group] or "" }
   end)
   return { content = M.table(lines) }
 end