+++ lua/lazy/view/render.lua
@@ -144,8 +144,6 @@ function M:title()
       if mode.name == "home" then
         if self.view.state.mode == "home" then
           title = " lazy.nvim " .. Config.options.ui.icons.lazy
-        else
-          title = " lazy.nvim (H) "
         end
       end
 
@@ -761,7 +759,7 @@ function M:debug()
         ---@type string[]
         plugins = vim.tbl_values(plugins)
         table.sort(plugins)
-        self:append("● ", "LazySpecial", { indent = 2 })
+        self:append(Config.options.ui.icons.debug, "LazySpecial", { indent = 2 })
         if handler_type == "keys" then
           for k, v in pairs(Config.plugins[plugins[1]]._.handlers.keys) do
             if k == value then